remove 'Unknown' as fallback name
It's better for the animations because otherwise our fake actor would also need to have 'Unknown' written on him. Besides, the main reason we needed it up till now was that it was necessary for the click-area when you want to edit the actor name. But the in-place-editing was removed.