remove unnecessary double selections (mousedown and click)
we already select on mousedown (dragStart), no need to select again on click. But a listener on click is needed to prevent the click from bubbling to the actor_editor, which would unselect the view.
Up till now the event chain was like this:
select -> unselect -> select (in the fraction of a second)
On 'mousedown' it would select, then when releasing the mouse 'click' gets fired which would propagate to the editor view and unselect it there, and finally the 'click'-handler on draggable would select the element again.
Now its just one select.